refactor tabs
